Split off nojs form
